Fix typing for time attribute in Data class

+++ python-setup/simcoon/data.py
@@ -16,7 +16,7 @@ class Data:
     def __init__(self, control: npt.NDArray[np.float_],
                  observation: npt.NDArray[np.float_],
                  time: Optional[npt.NDArray[np.float_]],
-                 timestep: Optional[np.float] = 0.01):
+                 timestep: Optional[float] = 0.01):
 
         self.control = control
         self.observation = observation
